+++ .gitignore
@@ -7,7 +7,8 @@
 node_modules
 
 parts
-snap
+stage
+prime
 
 *.snap
 *.tar.bz2
